+++ napi/parser/package.json
@@ -45,10 +45,10 @@
     "generated/deserialize/js_range.js",
     "generated/deserialize/ts_range.js",
     "generated/lazy/constructors.js",
-    "generated/lazy/types.js",
+    "generated/lazy/type_ids.js",
     "generated/lazy/walk.js",
     "generated/visit/keys.js",
-    "generated/visit/types.js",
+    "generated/visit/type_ids.js",
     "generated/visit/visitor.d.ts",
     "generated/visit/walk.js",
     "src-js/bindings.js",
